YQ-3617: fix GROUP BY HOP + AS_TABLE (#9370)
